+++ attmap/__init__.py
@@ -2,6 +2,7 @@
 
 from .attmap import AttMap
 from .attmap_echo import AttMapEcho
+from .helpers import *
 from .ordattmap import OrdAttMap
 from .ordpathex_attmap import OrdPathExAttMap
 from ._version import __version__
@@ -10,4 +11,4 @@ AttributeDict = AttMap
 AttributeDictEcho = AttMapEcho
 
 __all__ = ["AttMap", "AttMapEcho", "AttributeDict", "AttributeDictEcho",
-           "OrdAttMap", "OrdPathExAttMap"]
+           "OrdAttMap", "OrdPathExAttMap", "get_data_lines", "is_custom_map"]